# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ERP System API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /super-admin/businesses:
    get:
      operationId: listAllBusinesses
      tags: [super-admin]
      summary: List all businesses
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of businesses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessListResponse"
  /super-admin/businesses/{id}:
    get:
      operationId: getBusinessById
      tags: [super-admin]
      summary: Get business details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Business details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
    patch:
      operationId: updateBusinessStatus
      tags: [super-admin]
      summary: Update business status (activate/deactivate)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBusinessStatusBody"
      responses:
        "200":
          description: Business updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
  /super-admin/plans:
    get:
      operationId: listPlans
      tags: [super-admin]
      summary: List all subscription plans
      responses:
        "200":
          description: List of plans
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Plan"
                required: [data]
    post:
      operationId: createPlan
      tags: [super-admin]
      summary: Create a plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlanBody"
      responses:
        "201":
          description: Plan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users in current business
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
                required: [data]
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # PARTIES
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payments
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [receipt, payment]
        - name: partyId
          in: query
          schema:
            type: integer
        - name: fromDate
          in: query
          schema:
            type: string
        - name: toDate
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentListResponse"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Create payment (bill-wise or on-account)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentBody"
      responses:
        "201":
          description: Payment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get payment details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
    delete:
      operationId: deletePayment
      tags: [payments]
      summary: Delete payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /payments/outstanding:
    get:
      operationId: getOutstandingBills
      tags: [payments]
      summary: Get outstanding bills for a party
      parameters:
        - name: partyId
          in: query
          required: true
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
            enum: [receivable, payable]
      responses:
        "200":
          description: Outstanding bills
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutstandingResponse"

  # INVENTORY
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Plan:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        billingCycle:
          type: string
          enum: [monthly, yearly]
        maxUsers:
          type: integer
        features:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, name, price, billingCycle, isActive]
    CreatePlanBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        billingCycle:
          type: string
          enum: [monthly, yearly]
        maxUsers:
          type: integer
        features:
          type: array
          items:
            type: string
      required: [name, price, billingCycle]
    Business:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        businessCode:
          type: string
        gstin:
          type: string
        pan:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        stateCode:
          type: string
        pincode:
          type: string
        phone:
          type: string
        email:
          type: string
        businessType:
          type: string
        logo:
          type: string
        planId:
          type: integer
        planName:
          type: string
        status:
          type: string
          enum: [active, inactive, suspended]
        financialYearStart:
          type: string
        currency:
          type: string
        createdAt:
          type: string
        userCount:
          type: integer
      required: [id, name, businessCode, status]
    BusinessListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Business"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total]
...
    UpdateBusinessStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [active, inactive, suspended]
        planId:
          type: integer
      required: [status]
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [business_admin, staff]
        permissions:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, name, email, role, isActive]
    CreateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [business_admin, staff]
        permissions:
          type: array
          items:
            type: string
      required: [name, email, password, role]
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        permissions:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        password:
          type: string
    Party:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [customer, supplier, both]
        gstin:
          type: string
        pan:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        stateCode:
          type: string
        pincode:
          type: string
        openingBalance:
          type: number
        openingBalanceType:
          type: string
          enum: [debit, credit]
        creditLimit:
          type: number
        creditDays:
          type: integer
        isActive:
          type: boolean
        currentBalance:
          type: number
        customFields:
          type: object
        createdAt:
          type: string
      required: [id, name, type, isActive]
...
    Payment:
      type: object
      properties:
        id:
          type: integer
        paymentNumber:
          type: string
        type:
          type: string
          enum: [receipt, payment]
        date:
          type: string
        partyId:
          type: integer
        partyName:
          type: string
        amount:
          type: number
        paymentMode:
          type: string
          enum: [cash, bank, cheque, upi, other]
        referenceNumber:
          type: string
        notes:
          type: string
        isOnAccount:
          type: boolean
        allocations:
          type: array
          items:
            $ref: "#/components/schemas/PaymentAllocation"
        createdAt:
          type: string
      required: [id, type, date, partyId, amount, isOnAccount]
    PaymentAllocation:
      type: object
      properties:
        voucherId:
          type: integer
        voucherNumber:
          type: string
        voucherDate:
          type: string
        voucherAmount:
          type: number
        allocatedAmount:
          type: number
      required: [voucherId, allocatedAmount]
    CreatePaymentBody:
      type: object
      properties:
        type:
          type: string
          enum: [receipt, payment]
        date:
          type: string
        partyId:
          type: integer
        amount:
          type: number
        paymentMode:
          type: string
          enum: [cash, bank, cheque, upi, other]
        referenceNumber:
          type: string
        notes:
          type: string
        isOnAccount:
          type: boolean
        allocations:
          type: array
          items:
            type: object
            properties:
              voucherId:
                type: integer
              allocatedAmount:
                type: number
            required: [voucherId, allocatedAmount]
      required: [type, date, partyId, amount, paymentMode]
    PaymentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalAmount:
          type: number
      required: [data, total]
    OutstandingBill:
      type: object
      properties:
        voucherId:
          type: integer
        voucherNumber:
          type: string
        date:
          type: string
        dueDate:
          type: string
        originalAmount:
          type: number
        paidAmount:
          type: number
        balanceDue:
          type: number
        daysOverdue:
          type: integer
      required: [voucherId, voucherNumber, originalAmount, balanceDue]
    OutstandingResponse:
      type: object
      properties:
        party:
          $ref: "#/components/schemas/Party"
        totalOutstanding:
          type: number
        bills:
          type: array
          items:
            $ref: "#/components/schemas/OutstandingBill"
      required: [totalOutstanding, bills]
...